chore: Remove old `Settings` component and routes (#3409)
